1.20.1 Backport WIP

+++ src/main/java/com/leclowndu93150/wakes/event/WakeClientTicker.java
@@ -2,15 +2,16 @@ package com.leclowndu93150.wakes.event;
 
 import com.leclowndu93150.wakes.simulation.WakeHandler;
 import net.minecraft.client.Minecraft;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.client.event.ClientTickEvent;
+import net.minecraftforge.event.TickEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
 
-@EventBusSubscriber(bus =  EventBusSubscriber.Bus.GAME)
+@Mod.EventBusSubscriber(bus =  Mod.EventBusSubscriber.Bus.FORGE)
 public class WakeClientTicker {
     @SubscribeEvent
-    public static void onClientTick(ClientTickEvent.Pre event) {
+    public static void onClientTick(TickEvent.ClientTickEvent event) {
         Minecraft client = Minecraft.getInstance();
+        if(event.phase != TickEvent.Phase.START) return;
         if (client.level == null) {
             WakeHandler.kill();
         } else if (WakeHandler.getInstance().isEmpty()) {
